+++ lisp/eab-ui.el
@@ -89,6 +89,7 @@
 (setq require-final-newline nil)
 (setq mode-require-final-newline nil)
 (setq indent-tabs-mode nil)
+(add-hook 'org-mode-hook (lambda () (setq indent-tabs-mode nil)))
 (add-hook 'sh-mode-hook (lambda () (setq indent-tabs-mode nil)))
 (add-hook 'go-mode-hook (lambda ()
 			  (setq indent-tabs-mode nil)